% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdmean.R
\docType{data}
\name{nlswork}
\alias{nlswork}
\title{Data on wages and unemployment over time. From Stata: 
please see (and cite) \url{https://www.stata-press.com/data/r16/}}
\format{
A data frame with 18858 rows and 5 variables. Complete
observations from the original data, owned and released by Stata:
\describe{
  \item{ln_wage}{Natural log of wage}
  \item{idcode}{Respondent idcode}
  \item{year}{Year}
  \item{wks_work}{Amount worked}
  \item{union}{Union member (dummy)}
}
}
\source{
\url{https://www.stata-press.com/data/r16/}
}
\usage{
data(nlswork)
}
\description{
Data on wages and unemployment over time. From Stata: 
please see (and cite) \url{https://www.stata-press.com/data/r16/}
}
\keyword{datasets}
