% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdmean.R
\name{qdmean}
\alias{qdmean}
\title{Quasi-demean a variable, based on estimate of theta}
\usage{
qdmean(model, predictor, group, data, dv)
}
\arguments{
\item{model}{an estimated random effects model}

\item{predictor}{the predictor variable to quasi-demean}

\item{group}{the grouping variable to use to quasi-demean}

\item{data}{if a lmer model, the dataset used to estimate the model}

\item{dv}{if a lmer model, the dependent variable (as a string in quotes)}
}
\value{
the quasi-demeaned independent variable
}
\description{
Quasi-demean a variable, based on estimate of theta
}
\details{
if using a \code{lmer} model, pass the data with no \code{NA} observations
}
\examples{
\dontrun{
# using plm
model.plm <- plm(ln_wage ~ wks_work + union, index = c("idcode", "year"),
         data = nlswork,
         model = "random", random.method = "swar", effect = "individual")
# quasi-demean wks_work
wkswork.plm <- qdmean(model = model.plm, predictor = "wks_work", group = "idcode")
summary(wkswork.plm)

# using lmer
model.lmer <- lmer(ln_wage ~ wks_work + union + (1 | idcode),
         data = data)
# quasi-demean wks_work
wkswork.lmer <- qdmean(model = model.lmer, predictor = "wks_work", group = "idcode", 
         data = nlswork, dv = "ln_wage")
summary(wkswork.lmer)
}
}
\author{
Soren Jordan and Andrew Q. Philips
}
\keyword{demean}
